Refactor Footer into smaller, focused modules
Extracted social icons and contact dialog into separate modules as
requested in code review to improve maintainability and reduce the
size of the main Footer component file.

**Changes:**
- Created `Footer/SocialIcons.tsx` (194 lines): Social media icons,
  constants, and SocialDirectory component
- Created `Footer/ContactDialog.tsx` (68 lines): ContactDialog component
  and useContactDialog hook
- Refactored `Footer/index.tsx` from 561 to 315 lines (44% reduction)
- Maintained backward compatibility by re-exporting ContactDialog and
  useContactDialog from index.tsx
- All existing imports continue to work (no breaking changes)

**Benefits:**
- Improved code organization and separation of concerns
- Each module now has a single, focused responsibility
- Easier to navigate and understand the Footer component
- Facilitates future maintenance and testing

Addresses Review #26


Cleanup

Don't export icons

Footer is browser-only
